Initial testing.
Add-in the delete project.

Fix-up existing wait function; technically it is still a waituntil using should statement.

iniital commit of re-factor.

Re-factoring a bit of the code plus a testing demo.

latest commit.

Fix-up minor comments and hard-coded variables.

added stabilizing code (#24)

* added stabilizing code

* minor

Updating the tests to use beforeeach for clean up.

Add-in more stabilizing code for iframe and mouse events; there is an issue of mouse event on pseudo-elements.