BACKLOG-15722 sitemap panel UI first draft.
Updating the components to use moonstone plus fix-up css and dependencies.

Updating the scss plus some text improvements.

Updating some css and text formating. leverage theme from v4+.

Minor css touch up and margin fixes.

Minor fix on a hardcoded test variable.

Re-factor out the gql generic functions out.

Re-factor to use functional instead of classes.

Leverage the existing header component; fix-up some typography.

Add in the activate/save button plus the flush and submit button.

Add-in the disable functionality for indexURL.

Add-in the dialog functionality for the flush and submit.

Add-in the chip for sitemap index URL for required.

Add-in the mix-in for the properties on virtual-side node; re-factor to use formik.

Working version of the sitemap file section.

Add-in the snackbar portion; plus clean-up on old obsolete codes.

Add-in the missing compose.js

Add-in the query that is lost in rebase.